MA-45: Rename movableUC to nodeUC & moved node specific code to nodeUC
