+++ Messages/ReleaseNodeEdgeMessage.cs
@@ -2,10 +2,10 @@
 using mystery_app.ViewModels;
 
 namespace mystery_app.Messages;
-public class ReleaseNodeEdgeMessage : ValueChangedMessage<NodeViewModel>
+public class ReleaseNodeEdgeMessage : ValueChangedMessage<NodeViewModelBase>
 {
     // Can change from string to a custom node class
-    public ReleaseNodeEdgeMessage(NodeViewModel value) : base(value)
+    public ReleaseNodeEdgeMessage(NodeViewModelBase value) : base(value)
     {
     }
 }
