+++ Messages/SelectNodeEdgeMessage.cs
@@ -2,10 +2,10 @@
 using mystery_app.ViewModels;
 
 namespace mystery_app.Messages;
-public class SelectNodeEdgeMessage : ValueChangedMessage<NodeViewModel>
+public class SelectNodeEdgeMessage : ValueChangedMessage<NodeViewModelBase>
 {
     // TODO Can change from string to a custom node class
-    public SelectNodeEdgeMessage(NodeViewModel value) : base(value)
+    public SelectNodeEdgeMessage(NodeViewModelBase value) : base(value)
     {
     }
 }
