+++ Models/Edge.cs
@@ -5,23 +5,23 @@ namespace mystery_app.Models;
 
 public class Edge
 {
-    public Edge(NodeViewModel fromNode, NodeViewModel toNode, string description="")
+    public Edge(NodeViewModelBase fromNode, NodeViewModelBase toNode, string description="")
     {
         FromNode = fromNode;
         ToNode = toNode;
         Description = description;
     }
 
-    private NodeViewModel _fromNode;
-    private NodeViewModel _toNode;
+    private NodeViewModelBase _fromNode;
+    private NodeViewModelBase _toNode;
     private string _description;
 
-    public NodeViewModel FromNode
+    public NodeViewModelBase FromNode
     {
         get { return _fromNode; }
         set { _fromNode = value; }
     }
-    public NodeViewModel ToNode
+    public NodeViewModelBase ToNode
     {
         get { return _toNode; }
         set { _toNode = value; }
