+++ Views/NodeView.axaml
@@ -1,4 +1,4 @@
-<controls:MovableUserControl
+<controls:NodeUserControl
 	xmlns="https://github.com/avaloniaui"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -58,4 +58,4 @@
 			PointerPressed="SelectResize"
 			PointerReleased="ReleaseResize"/>
 	</DockPanel>
-</controls:MovableUserControl>
+</controls:NodeUserControl>
